requirements.txt
flask
python-dotenv
requests
beautifulsoup4
pandas
numpy
scikit-learn
joblib
lxml
